refactor: consolidate dashboard/ into dashboard-v2/
Move all dashboard engine files (DashboardEngine, DashboardRenderer,
SessionPanel, zones/) into dashboard-v2/ and update all imports.
The old dashboard/ directory is removed.
