+++ apps/web/src/components/dashboard-v2/SessionPanel.tsx
