+++ apps/web/src/components/dashboard-v2/SessionPanelCollapsed.tsx
