+++ apps/web/src/components/dashboard-v2/__tests__/SessionPanel.test.tsx
