+++ apps/web/src/components/dashboard-v2/useSessionSlot.ts
