+++ apps/web/src/components/dashboard-v2/zones/AgentsSidebar.tsx
