+++ apps/web/src/components/dashboard-v2/zones/HeroZone.tsx
