+++ apps/web/src/components/dashboard-v2/zones/StatsZone.tsx
