+++ apps/web/src/components/dashboard-v2/zones/__tests__/AgentsSidebar.test.tsx
