+++ apps/web/src/components/dashboard-v2/zones/__tests__/HeroZone.test.tsx
