+++ apps/web/src/components/dashboard-v2/zones/__tests__/ScoreboardZone.test.tsx
