+++ apps/web/src/components/dashboard-v2/zones/__tests__/SessionBar.test.tsx
