+++ apps/web/src/components/dashboard-v2/zones/__tests__/StatsZone.test.tsx
