--- apps/web/src/components/dashboard/index.ts
@@ -1,5 +1,0 @@
-export { dashboardMachine } from './DashboardEngine';
-export type { DashboardEvent, DashboardEngineContext } from './DashboardEngine';
-export { DashboardEngineProvider } from './DashboardEngineProvider';
-export { DashboardRenderer } from './DashboardRenderer';
-export { useDashboardMode } from './useDashboardMode';
